+++ dd-java-agent/instrumentation/servlet-3/src/main/java/datadog/trace/instrumentation/servlet3/AbstractServlet3Instrumentation.java
@@ -1,5 +1,6 @@
 package datadog.trace.instrumentation.servlet3;
 
+import static datadog.trace.agent.tooling.ClassLoaderMatcher.classLoaderHasClasses;
 import static net.bytebuddy.matcher.ElementMatchers.failSafe;
 import static net.bytebuddy.matcher.ElementMatchers.hasSuperType;
 import static net.bytebuddy.matcher.ElementMatchers.isInterface;
@@ -8,18 +9,34 @@ import static net.bytebuddy.matcher.ElementMatchers.named;
 import static net.bytebuddy.matcher.ElementMatchers.not;
 import static net.bytebuddy.matcher.ElementMatchers.takesArgument;
 
-import com.google.auto.service.AutoService;
 import datadog.trace.agent.tooling.Instrumenter;
 import java.util.HashMap;
 import java.util.Map;
 import net.bytebuddy.matcher.ElementMatcher;
 
-@AutoService(Instrumenter.class)
-public final class FilterChain3Instrumentation extends AbstractServlet3Instrumentation {
+public class AbstractServlet3Instrumentation extends Instrumenter.Default {
+
+  public AbstractServlet3Instrumentation() {
+    super("servlet", "servlet-3");
+  }
 
   @Override
   public ElementMatcher typeMatcher() {
     return not(isInterface()).and(failSafe(hasSuperType(named("javax.servlet.FilterChain"))));
+  }
+
+  @Override
+  public ElementMatcher<? super ClassLoader> classLoaderMatcher() {
+    return classLoaderHasClasses("javax.servlet.AsyncEvent", "javax.servlet.AsyncListener");
+  }
+
+  @Override
+  public String[] helperClassNames() {
+    return new String[] {
+      "datadog.trace.instrumentation.servlet3.HttpServletRequestExtractAdapter",
+      "datadog.trace.instrumentation.servlet3.HttpServletRequestExtractAdapter$MultivaluedMapFlatIterator",
+      "datadog.trace.instrumentation.servlet3.TagSettingAsyncListener"
+    };
   }
 
   @Override
